Move breakpoint definition and comparison utils
